Added .Rmd for the generation of the heatmap of Treg signature genes (Figure 1 f).
Minor edits to .Rmd for TregDGEScatterplot for consistency of path descriptions.

Readme.md was updated for general usage of the .Rmd files for bulkRNAseq_analyses.

Also, the build for R (RStudio, R and all dependencies) on the machine where the code was developed were restored from 07.10.2021 via time machine backup because of a reproducibility issue with Figure 1 f. This issues could not be resolved. We proceed with the data strucutre as is.