Figure 3 and 4 related UMAP, Heatmaps, GSEA and Gene Expression Barplots.
All figures related to Figure 3 and 4 as well as respective supplementary figures based on the complete set of bulkRNAseq samples (pcbd/pro.prc.bas.don.) are covered in this .Rmd file.

Unsure about the consitency check with read-in Graphia .tsv file: After merging one gene is missing which is still pointed out in this .Rmd even though it is hardly disturbing the analysis.